refactor(rbac): move workflow_rbac.py to services/ and router to feature_routers (#2323)
Move workflow_rbac.py from backend root to services/ for consistency with
co-located service modules. Move workflow_permissions router registration
from core_routers (hard fail on import error) to feature_routers (graceful
degradation via importlib), matching the pattern used by api.permissions.
